projects/movie_review/requirements.txt
asgiref==3.2.10
beautifulsoup4==4.9.3
Django==3.1.2
django-appconf==1.0.4
django-bootstrap4==2.2.0
django-imagekit==4.0.2
importlib-metadata==1.7.0
pilkit==2.0
Pillow==7.2.0
pytz==2020.1
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
zipp==3.3.0
